requirements.txt
numpy
pillow
affine
shapely
tqdm
